+++ helper/widgets/html.js
@@ -1,3 +1,5 @@
+const EDITOR = Symbol('editor');
+
 /**
  * The Widget HTML class
  *
@@ -24,6 +26,54 @@ Html.constitute(function prepareSchema() {
 });
 
 /**
+ * Start the editor
+ *
+ * @author   Jelle De Loecker   <jelle@elevenways.be>
+ * @since    0.2.1
+ * @version  0.2.1
+ */
+Html.setMethod(async function _startEditor() {
+
+	if (this[EDITOR]) {
+		this[EDITOR].destroy();
+	}
+
+	let ckeditor_path = hawkejs.scene.exposed.ckeditor_path;
+
+	if (!ckeditor_path) {
+		return;
+	}
+
+	await hawkejs.require(ckeditor_path);
+
+	const options = {
+		toolbar: hawkejs.scene.exposed.ckeditor_toolbar,
+		updateSourceElementOnDestroy: true,
+		simpleUpload: {
+			uploadUrl: '/api/block/upload',
+			withCredentials: true,
+		},
+	};
+
+	let editor = await InlineEditor.create(this.widget, options);
+	this[EDITOR] = editor;
+});
+
+/**
+ * Stop the editor
+ *
+ * @author   Jelle De Loecker   <jelle@elevenways.be>
+ * @since    0.2.1
+ * @version  0.2.1
+ */
+Html.setMethod(function _stopEditor() {
+	if (this[EDITOR]) {
+		this[EDITOR].destroy();
+		this[EDITOR] = null;
+	}
+});
+
+/**
  * Populate the widget
  *
  * @author   Jelle De Loecker   <jelle@elevenways.be>
@@ -34,7 +84,6 @@ Html.constitute(function prepareSchema() {
  */
 Html.setMethod(function populateWidget() {
 	this.widget.innerHTML = this.config.html;
-
 	populateWidget.super.call(this);
 });