+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.1 (WIP)
 
 * Add a `-first` modifier class to the first visible entry of a table-of-contents element
+* Make HTML widget editable
 
 ## 0.2.0 (2022-11-02)